requirements.txt
schedule~=1.2.2
atproto~=0.0.57
requests~=2.32.3
pillow~=10.4.0